Refactor and create object for filePath
