+++ src/test/java/org/fungover/storm/fileHandler/FileRequestHandlerTest.java
@@ -1,12 +1,11 @@
 package org.fungover.storm.fileHandler;
 
+import org.fungover.storm.fileHandler.re.FileRequestHandler;
 import org.junit.jupiter.api.Test;
 
-import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 
 import static org.assertj.core.api.Assertions.assertThat;
